+++ lib/src/dependencies/path.dart
@@ -87,8 +87,23 @@ extension ExtensionPathString on String {
     String? part13,
     String? part14,
     String? part15,
-  ]) =>
-      path.absolute(this, part2, part3, part4, part5, part6, part7, part8, part9, part10, part11, part12, part13, part14, part15);
+  ]) => path.absolute(
+    this,
+    part2,
+    part3,
+    part4,
+    part5,
+    part6,
+    part7,
+    part8,
+    part9,
+    part10,
+    part11,
+    part12,
+    part13,
+    part14,
+    part15,
+  );
 }
 
 extension ExtensionPathIterableString on Iterable<String> {