+++ lib/src/dependencies/plugin/path_provider.dart
@@ -11,10 +11,12 @@ class PathProvider {
   static PathProvider? _singleton;
 
   /// Android	iOS	Linux	macOS	Windows
-  Future<Directory?> getTemporaryDirectory() async => isWeb ? null : await path.getTemporaryDirectory();
+  Future<Directory?> getTemporaryDirectory() async =>
+      isWeb ? null : await path.getTemporaryDirectory();
 
   /// Android	iOS	Linux	macOS	Windows
-  Future<Directory?> getApplicationSupportDirectory() async => isWeb ? null : await path.getApplicationSupportDirectory();
+  Future<Directory?> getApplicationSupportDirectory() async =>
+      isWeb ? null : await path.getApplicationSupportDirectory();
 
   /// Android	iOS	Linux	macOS	Windows
   /// Android for [getExternalStorageDirectory()]
@@ -25,13 +27,16 @@ class PathProvider {
   }
 
   ///	iOS	macOS
-  Future<Directory?> getLibraryDirectory() async => isWeb ? null : await path.getLibraryDirectory();
+  Future<Directory?> getLibraryDirectory() async =>
+      isWeb ? null : await path.getLibraryDirectory();
 
   /// Android	iOS	Linux	macOS	Windows
-  Future<Directory?> getApplicationDocumentsDirectory() async => isWeb ? null : await path.getApplicationDocumentsDirectory();
+  Future<Directory?> getApplicationDocumentsDirectory() async =>
+      isWeb ? null : await path.getApplicationDocumentsDirectory();
 
   /// Android	iOS	Linux	macOS	Windows
-  Future<Directory?> getApplicationCacheDirectory() async => isWeb ? null : await path.getApplicationCacheDirectory();
+  Future<Directory?> getApplicationCacheDirectory() async =>
+      isWeb ? null : await path.getApplicationCacheDirectory();
 
   /// Android	iOS	Linux	macOS	Windows
   /// Android for [getExternalCacheDirectories()]
@@ -42,16 +47,22 @@ class PathProvider {
   }
 
   /// Android iOS	Linux	macOS	Windows
-  Future<Directory?> getDownloadsDirectory() async => isWeb ? null : await path.getDownloadsDirectory();
+  Future<Directory?> getDownloadsDirectory() async =>
+      isWeb ? null : await path.getDownloadsDirectory();
 
   /// Android
-  Future<Directory?> getExternalStorageDirectory() async => isWeb || !isAndroid ? null : await path.getExternalStorageDirectory();
+  Future<Directory?> getExternalStorageDirectory() async =>
+      isWeb || !isAndroid ? null : await path.getExternalStorageDirectory();
 
   /// Android
   Future<List<Directory>?> getExternalCacheDirectories() async =>
       isWeb || !isAndroid ? null : await path.getExternalCacheDirectories();
 
   /// Android
-  Future<List<Directory>?> getExternalStorageDirectories({StorageDirectory? type}) async =>
-      isWeb || !isAndroid ? null : await path.getExternalStorageDirectories(type: type);
+  Future<List<Directory>?> getExternalStorageDirectories({
+    StorageDirectory? type,
+  }) async =>
+      isWeb || !isAndroid
+          ? null
+          : await path.getExternalStorageDirectories(type: type);
 }